# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CRM2 API",
    "version": "2.0.0",
    "description": "Generated from the live Express app (ADR-0031, B-11). Phase 1 = surface contract (paths/methods/tags/security). Request/response schemas land in phase 2 (zod single source). Do not edit by hand — run `pnpm openapi`."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v2/billing/cases": {
      "get": {
        "tags": [
          "billing"
        ],
        "operationId": "get_api_v2_billing_cases",
        "summary": "GET /api/v2/billing/cases",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/cases": {
      "get": {
        "tags": [
          "cases"
        ],
        "operationId": "get_api_v2_cases",
        "summary": "GET /api/v2/cases",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "cases"
        ],
        "operationId": "post_api_v2_cases",
        "summary": "POST /api/v2/cases",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/client-products": {
      "get": {
        "tags": [
          "client-products"
        ],
        "operationId": "get_api_v2_client_products",
        "summary": "GET /api/v2/client-products",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "client-products"
        ],
        "operationId": "post_api_v2_client_products",
        "summary": "POST /api/v2/client-products",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/clients": {
      "get": {
        "tags": [
          "clients"
        ],
        "operationId": "get_api_v2_clients",
        "summary": "GET /api/v2/clients",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "clients"
        ],
        "operationId": "post_api_v2_clients",
        "summary": "POST /api/v2/clients",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/commission-rates": {
      "get": {
        "tags": [
          "commission-rates"
        ],
        "operationId": "get_api_v2_commission_rates",
        "summary": "GET /api/v2/commission-rates",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "commission-rates"
        ],
        "operationId": "post_api_v2_commission_rates",
        "summary": "POST /api/v2/commission-rates",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/cpv-units": {
      "get": {
        "tags": [
          "cpv-units"
        ],
        "operationId": "get_api_v2_cpv_units",
        "summary": "GET /api/v2/cpv-units",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "cpv-units"
        ],
        "operationId": "post_api_v2_cpv_units",
        "summary": "POST /api/v2/cpv-units",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/departments": {
      "get": {
        "tags": [
          "departments"
        ],
        "operationId": "get_api_v2_departments",
        "summary": "GET /api/v2/departments",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "departments"
        ],
        "operationId": "post_api_v2_departments",
        "summary": "POST /api/v2/departments",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/designations": {
      "get": {
        "tags": [
          "designations"
        ],
        "operationId": "get_api_v2_designations",
        "summary": "GET /api/v2/designations",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "designations"
        ],
        "operationId": "post_api_v2_designations",
        "summary": "POST /api/v2/designations",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "operationId": "get_api_v2_jobs",
        "summary": "GET /api/v2/jobs",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/locations": {
      "get": {
        "tags": [
          "locations"
        ],
        "operationId": "get_api_v2_locations",
        "summary": "GET /api/v2/locations",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "locations"
        ],
        "operationId": "post_api_v2_locations",
        "summary": "POST /api/v2/locations",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/notifications": {
      "delete": {
        "tags": [
          "notifications"
        ],
        "operationId": "delete_api_v2_notifications",
        "summary": "DELETE /api/v2/notifications",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "notifications"
        ],
        "operationId": "get_api_v2_notifications",
        "summary": "GET /api/v2/notifications",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/policies": {
      "get": {
        "tags": [
          "policies"
        ],
        "operationId": "get_api_v2_policies",
        "summary": "GET /api/v2/policies",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "policies"
        ],
        "operationId": "post_api_v2_policies",
        "summary": "POST /api/v2/policies",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/products": {
      "get": {
        "tags": [
          "products"
        ],
        "operationId": "get_api_v2_products",
        "summary": "GET /api/v2/products",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "products"
        ],
        "operationId": "post_api_v2_products",
        "summary": "POST /api/v2/products",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/rate-types": {
      "get": {
        "tags": [
          "rate-types"
        ],
        "operationId": "get_api_v2_rate_types",
        "summary": "GET /api/v2/rate-types",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v2/rates": {
      "get": {
        "tags": [
          "rates"
        ],
        "operationId": "get_api_v2_rates",
        "summary": "GET /api/v2/rates",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "rates"
        ],
        "operationId": "post_api_v2_rates",
        "summary": "POST /api/v2/rates",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/report-layouts": {
      "get": {
        "tags": [
          "report-layouts"
        ],
        "operationId": "get_api_v2_report_layouts",
        "summary": "GET /api/v2/report-layouts",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "report-layouts"
        ],
        "operationId": "post_api_v2_report_layouts",
        "summary": "POST /api/v2/report-layouts",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/report-templates": {
      "get": {
        "tags": [
          "report-templates"
        ],
        "operationId": "get_api_v2_report_templates",
        "summary": "GET /api/v2/report-templates",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "report-templates"
        ],
        "operationId": "post_api_v2_report_templates",
        "summary": "POST /api/v2/report-templates",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/roles": {
      "get": {
        "tags": [
          "roles"
        ],
        "operationId": "get_api_v2_roles",
        "summary": "GET /api/v2/roles",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "roles"
        ],
        "operationId": "post_api_v2_roles",
        "summary": "POST /api/v2/roles",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/saved-views": {
      "get": {
        "tags": [
          "saved-views"
        ],
        "operationId": "get_api_v2_saved_views",
        "summary": "GET /api/v2/saved-views",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "saved-views"
        ],
        "operationId": "post_api_v2_saved_views",
        "summary": "POST /api/v2/saved-views",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "operationId": "get_api_v2_tasks",
        "summary": "GET /api/v2/tasks",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/users": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get_api_v2_users",
        "summary": "GET /api/v2/users",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "post_api_v2_users",
        "summary": "POST /api/v2/users",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/verification-units": {
      "get": {
        "tags": [
          "verification-units"
        ],
        "operationId": "get_api_v2_verification_units",
        "summary": "GET /api/v2/verification-units",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "verification-units"
        ],
        "operationId": "post_api_v2_verification_units",
        "summary": "POST /api/v2/verification-units",
        "responses": {
          "default": {
            "description": "See @crm2/sdk for the typed shape (response schemas land in phase 2)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}